+++ packages/js-plugin-nft-storage/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @metaplex-foundation/js-plugin-nft-storage
 
+## 0.18.3
+
+### Patch Changes
+
+- Updated dependencies [b6d9571f]
+  - @metaplex-foundation/js@0.18.3
+
+## 0.18.2
+
+### Patch Changes
+
+- Updated dependencies [10644449]
+- Updated dependencies [2c01466e]
+  - @metaplex-foundation/js@0.18.2
+
 ## 0.18.1
 
 ### Patch Changes